+++ app/controllers/staff/budgets_controller.rb
@@ -9,7 +9,7 @@ class Staff::BudgetsController < Staff::BaseController
 
     authorize @budget
 
-    unless @activity.is_programme_level?
+    unless @activity.is_programme_level? | @activity.is_project_level?
       flash[:warning] = I18n.t("page_title.errors.budget.not_possible")
       redirect_to organisation_activity_path(@activity.organisation, @activity)
     end