+++ app/helpers/activity_helper.rb
@@ -13,6 +13,6 @@ module ActivityHelper
 
   def activity_back_path(activity)
     return organisation_path(activity.organisation) if activity.is_fund_level?
-    return organisation_activity_path(activity.parent_activity.organisation, activity.parent_activity) if activity.is_programme_level?
+    organisation_activity_path(activity.parent_activity.organisation, activity.parent_activity)
   end
 end